+++ app/models/issue.rb
@@ -17,8 +17,7 @@
 #
 
 class Issue < ActiveRecord::Base
-  include IssueCommonality
-  include Votes
+  include Issuable
 
   attr_accessible :title, :assignee_id, :closed, :position, :description,
                   :milestone_id, :label_list, :author_id_of_changes
