+++ app/observers/key_observer.rb
@@ -1,12 +1,12 @@
 class KeyObserver < ActiveRecord::Observer
-  include GitHost
+  include Gitolited
 
   def after_save(key)
-    git_host.set_key(key.identifier, key.key, key.projects)
+    gitolite.set_key(key.identifier, key.key, key.projects)
   end
 
   def after_destroy(key)
     return if key.is_deploy_key && !key.last_deploy?
-    git_host.remove_key(key.identifier, key.projects)
+    gitolite.remove_key(key.identifier, key.projects)
   end
 end
