--- lib/git_host.rb
@@ -1,11 +1,0 @@
-# == GitHost role
-#
-# Provide a shortcut to Gitlab::Gitolite instance
-#
-# Used by Project, UsersProject
-#
-module GitHost
-  def git_host
-    Gitlab::Gitolite.new
-  end
-end
